fix: refresh shortcut matcher synchronously
